+++ src/main/java/com/cgm/infolab/db/model/ChatMessageEntity.java
@@ -1,4 +1,6 @@
-package com.cgm.infolab.db;
+package com.cgm.infolab.db.model;
+
+import com.cgm.infolab.db.ID;
 
 import java.sql.Timestamp;
 import java.util.Objects;
