+++ src/main/java/com/cgm/infolab/db/model/UserEntity.java
@@ -1,16 +1,21 @@
-package com.cgm.infolab.db;
+package com.cgm.infolab.db.model;
 
-public class RoomEntity {
+import com.cgm.infolab.db.ID;
+
+public class UserEntity {
     private long id;
     private String name;
 
-    private RoomEntity(long id, String name) {
+    public UserEntity() {
+    }
+
+    private UserEntity(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public static RoomEntity of(String name) {
-        return new RoomEntity(ID.None, name);
+    public static UserEntity of(String name) {
+        return new UserEntity(ID.None, name);
     }
 
     public long getId() {
