+++ src/main/java/com/cgm/infolab/db/repository/ChatMessageRepository.java
@@ -1,5 +1,7 @@
-package com.cgm.infolab.db;
+package com.cgm.infolab.db.repository;
 
+import com.cgm.infolab.db.model.ChatMessageEntity;
+import com.cgm.infolab.db.model.RoomEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.DuplicateKeyException;
