+++ src/main/java/com/cgm/infolab/db/repository/RoomRepository.java
@@ -1,11 +1,8 @@
-package com.cgm.infolab.db;
+package com.cgm.infolab.db.repository;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.cgm.infolab.db.model.RoomEntity;
 import org.springframework.dao.DuplicateKeyException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Component;
