+++ src/main/java/com/cgm/infolab/db/repository/UserRepository.java
@@ -1,11 +1,7 @@
-package com.cgm.infolab.db;
+package com.cgm.infolab.db.repository;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.dao.EmptyResultDataAccessException;
+import com.cgm.infolab.db.model.UserEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Component;
